+++ ls_UI/core/changelog.lua
@@ -6,5 +6,10 @@ local _G = getfenv(0)
 
 -- Mine
 E.CHANGELOG = [[
-- Fixed fading. Adding new stuff is never easy...
+- Added 10.1.0 support.
+- Removed login message for good. It served its purpose.
+
+### Minimap
+
+- Fixed an issue where the difficulty flag's position wasn't adjusted when flipping the minimap.
 ]]