+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## Version 100100.01
+
+- Added 10.1.0 support.
+
+- Removed login message for good. It served its purpose.
+
+### Minimap
+
+- Fixed an issue where the difficulty flag's position wasn't adjusted when flipping the minimap.
+
 ## Version 100005.04
 
 - Fixed fading. Adding new stuff is never easy...